requirements.txt
streamlit==1.31.0
openai==1.12.0
anthropic==0.8.0
groq==0.4.0
google-generativeai==0.3.2
cohere==4.39
python-dotenv==1.0.0
networkx==3.2.1
plotly==5.18.0
pandas==2.2.0
